Created initial page with all filters
